refactor: 랭킹계산함수 분리
